FIR/UAST: commonize constructor declarations
