+++ plugins/uast-kotlin-fir/src/org/jetbrains/uast/kotlin/declarations/FirKotlinUMethod.kt
@@ -10,35 +10,25 @@ import org.jetbrains.kotlin.asJava.LightClassUtil
 import org.jetbrains.kotlin.asJava.elements.*
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.containingClassOrObject
-import org.jetbrains.kotlin.utils.SmartList
-import org.jetbrains.kotlin.utils.addToStdlib.firstIsInstanceOrNull
 import org.jetbrains.uast.*
 
-open class FirKotlinUMethod(
+class FirKotlinUMethod(
     psi: PsiMethod,
     sourcePsi: KtDeclaration?,
     givenParent: UElement?
-) : BaseKotlinUMethod(psi, sourcePsi, givenParent) {
+) : BaseKotlinUMethod(psi, sourcePsi, givenParent), FirKotlinUMethodParametersProducer {
     constructor(
         psi: KtLightMethod,
         givenParent: UElement?
     ) : this(psi, getKotlinMemberOrigin(psi), givenParent)
 
-    override val uastParameters: List<UParameter> by lz {
-        val lightParams = psi.parameterList.parameters
-        val receiverTypeReference =
-            receiverTypeReference ?: return@lz lightParams.map { FirKotlinUParameter(it, getKotlinMemberOrigin(it), this) }
-        val lightReceiver = lightParams.firstOrNull() ?: return@lz emptyList<UParameter>()
-        val uParameters = SmartList<UParameter>(FirKotlinReceiverUParameter(lightReceiver, receiverTypeReference, this))
-        lightParams.drop(1).mapTo(uParameters) { FirKotlinUParameter(it, getKotlinMemberOrigin(it), this) }
-        uParameters
-    }
+    override val uastParameters: List<UParameter> by lz { produceUastParameters(this, receiverTypeReference) }
 
     companion object {
         fun create(
             psi: KtLightMethod,
             givenParent: UElement?
-        ): FirKotlinUMethod {
+        ): UMethod {
             return when (psi) {
                 is KtConstructor<*> ->
                     FirKotlinConstructorUMethod(psi.containingClassOrObject, psi, givenParent)
@@ -51,7 +41,7 @@ open class FirKotlinUMethod(
         fun create(
             sourcePsi: KtDeclaration?,
             givenParent: UElement?
-        ): FirKotlinUMethod? {
+        ): UMethod? {
             val javaPsi = when (sourcePsi) {
                 is KtPropertyAccessor ->
                     LightClassUtil.getLightClassAccessorMethod(sourcePsi)
@@ -67,78 +57,5 @@ open class FirKotlinUMethod(
                     FirKotlinUMethod(javaPsi, sourcePsi, givenParent)
             }
         }
-    }
-}
-
-// TODO: can be commonized if *KotlinUMethod is commonized
-open class FirKotlinConstructorUMethod(
-    private val ktClass: KtClassOrObject?,
-    override val psi: PsiMethod,
-    kotlinOrigin: KtDeclaration?,
-    givenParent: UElement?
-) : FirKotlinUMethod(psi, kotlinOrigin, givenParent) {
-    constructor(
-        ktClass: KtClassOrObject?,
-        psi: KtLightMethod,
-        givenParent: UElement?
-    ) : this(ktClass, psi, psi.kotlinOrigin, givenParent)
-
-    override val javaPsi = psi
-
-    val isPrimary: Boolean
-        get() = sourcePsi is KtPrimaryConstructor || sourcePsi is KtClassOrObject
-
-    override val uastBody: UExpression? by lz {
-        val delegationCall: KtCallElement? = sourcePsi.let {
-            when {
-                isPrimary -> ktClass?.superTypeListEntries?.firstIsInstanceOrNull<KtSuperTypeCallEntry>()
-                it is KtSecondaryConstructor -> it.getDelegationCall()
-                else -> null
-            }
-        }
-        val bodyExpressions = getBodyExpressions()
-        if (delegationCall == null && bodyExpressions.isEmpty()) return@lz null
-        KotlinLazyUBlockExpression(this) { uastParent ->
-            SmartList<UExpression>().apply {
-                delegationCall?.let {
-                    // TODO: function call for delegationCall
-                    add(UastEmptyExpression(uastParent))
-                }
-                bodyExpressions.forEach {
-                    add(baseResolveProviderService.baseKotlinConverter.convertOrEmpty(it, uastParent))
-                }
-            }
-        }
-    }
-
-    override val uastAnchor: UIdentifier? by lz {
-        KotlinUIdentifier(
-            javaPsi.nameIdentifier,
-            if (isPrimary) ktClass?.nameIdentifier else (sourcePsi as? KtSecondaryConstructor)?.getConstructorKeyword(),
-            this
-        )
-    }
-
-    protected open fun getBodyExpressions(): List<KtExpression> {
-        if (isPrimary) return getInitializers()
-        val bodyExpression = (sourcePsi as? KtFunction)?.bodyExpression ?: return emptyList()
-        if (bodyExpression is KtBlockExpression) return bodyExpression.statements
-        return listOf(bodyExpression)
-    }
-
-    protected fun getInitializers(): List<KtExpression> {
-        return ktClass?.getAnonymousInitializers()?.mapNotNull { it.body } ?: emptyList()
-    }
-}
-
-// TODO: can be commonized if *KotlinUMethod is commonized
-//   also reuse the comments there (about KT-21617)
-class FirKotlinSecondaryConstructorWithInitializersUMethod(
-    ktClass: KtClassOrObject?,
-    psi: KtLightMethod,
-    givenParent: UElement?
-) : FirKotlinConstructorUMethod(ktClass, psi, givenParent) {
-    override fun getBodyExpressions(): List<KtExpression> {
-        return getInitializers() + super.getBodyExpressions()
     }
 }
