+++ plugins/uast-kotlin/src/org/jetbrains/uast/kotlin/declarations/KotlinUMethod.kt
@@ -21,41 +21,24 @@ import org.jetbrains.kotlin.asJava.elements.*
 import org.jetbrains.kotlin.psi.*
 import org.jetbrains.kotlin.psi.psiUtil.containingClassOrObject
 import org.jetbrains.kotlin.psi.psiUtil.getParentOfType
-import org.jetbrains.kotlin.utils.SmartList
-import org.jetbrains.kotlin.utils.addToStdlib.firstIsInstanceOrNull
 import org.jetbrains.uast.*
 import org.jetbrains.uast.kotlin.*
 import org.jetbrains.uast.kotlin.psi.UastFakeLightMethod
-import org.jetbrains.uast.kotlin.psi.UastKotlinPsiParameter
 
 open class KotlinUMethod(
     psi: PsiMethod,
     sourcePsi: KtDeclaration?,
     givenParent: UElement?
-) : BaseKotlinUMethod(psi, sourcePsi, givenParent) {
+) : BaseKotlinUMethod(psi, sourcePsi, givenParent), KotlinUMethodParametersProducer {
     constructor(
         psi: KtLightMethod,
         givenParent: UElement?
     ) : this(psi, getKotlinMemberOrigin(psi), givenParent)
 
-    override val uastParameters by lz {
-
-        fun parameterOrigin(psiParameter: PsiParameter?): KtElement? = when (psiParameter) {
-            is KtLightElement<*, *> -> psiParameter.kotlinOrigin
-            is UastKotlinPsiParameter -> psiParameter.ktParameter
-            else -> null
-        }
-
-        val lightParams = psi.parameterList.parameters
-        val receiver = receiverTypeReference ?: return@lz lightParams.map { KotlinUParameter(it, parameterOrigin(it), this) }
-        val receiverLight = lightParams.firstOrNull() ?: return@lz emptyList<UParameter>()
-        val uParameters = SmartList<UParameter>(KotlinReceiverUParameter(receiverLight, receiver, this))
-        lightParams.drop(1).mapTo(uParameters) { KotlinUParameter(it, parameterOrigin(it), this) }
-        uParameters
-    }
+    override val uastParameters by lz { produceUastParameters(this, receiverTypeReference) }
 
     companion object {
-        fun create(psi: KtLightMethod, containingElement: UElement?): KotlinUMethod {
+        fun create(psi: KtLightMethod, containingElement: UElement?): UMethod {
             val kotlinOrigin = psi.kotlinOrigin
             return if (kotlinOrigin is KtConstructor<*>) {
                 KotlinConstructorUMethod(
@@ -69,76 +52,6 @@ open class KotlinUMethod(
                 KotlinUMethod(psi, containingElement)
         }
     }
-}
-
-open class KotlinConstructorUMethod(
-    private val ktClass: KtClassOrObject?,
-    override val psi: PsiMethod,
-    kotlinOrigin: KtDeclaration?,
-    givenParent: UElement?
-) : KotlinUMethod(psi, kotlinOrigin, givenParent) {
-
-    constructor(
-        ktClass: KtClassOrObject?,
-        psi: KtLightMethod,
-        givenParent: UElement?
-    ) : this(ktClass, psi, psi.kotlinOrigin, givenParent)
-
-    val isPrimary: Boolean
-        get() = sourcePsi.let { it is KtPrimaryConstructor || it is KtClassOrObject }
-
-    override val uastBody: UExpression? by lz {
-        val delegationCall: KtCallElement? = sourcePsi.let {
-            when {
-                isPrimary -> ktClass?.superTypeListEntries?.firstIsInstanceOrNull<KtSuperTypeCallEntry>()
-                it is KtSecondaryConstructor -> it.getDelegationCall()
-                else -> null
-            }
-        }
-        val bodyExpressions = getBodyExpressions()
-        if (delegationCall == null && bodyExpressions.isEmpty()) return@lz null
-        KotlinLazyUBlockExpression(this) { uastParent ->
-            SmartList<UExpression>().apply {
-                delegationCall?.let {
-                    add(KotlinUFunctionCallExpression(it, uastParent))
-                }
-                bodyExpressions.forEach {
-                    add(KotlinConverter.convertOrEmpty(it, uastParent))
-                }
-            }
-        }
-    }
-
-    override val uastAnchor: KotlinUIdentifier by lz {
-        KotlinUIdentifier(
-            psi.nameIdentifier,
-            if (isPrimary) ktClass?.nameIdentifier else (sourcePsi as? KtSecondaryConstructor)?.getConstructorKeyword(),
-            this
-        )
-    }
-
-    override val javaPsi = psi
-
-    protected open fun getBodyExpressions(): List<KtExpression> {
-        if (isPrimary) return getInitializers()
-        val bodyExpression = (sourcePsi as? KtFunction)?.bodyExpression ?: return emptyList()
-        if (bodyExpression is KtBlockExpression) return bodyExpression.statements
-        return listOf(bodyExpression)
-    }
-
-    protected fun getInitializers() = ktClass?.getAnonymousInitializers()?.mapNotNull { it.body } ?: emptyList()
-
-}
-
-// This class was created as a workaround for KT-21617 to be the only constructor which includes `init` block
-// when there is no primary constructors in the class.
-// It is expected to have only one constructor of this type in a UClass.
-class KotlinSecondaryConstructorWithInitializersUMethod(
-    ktClass: KtClassOrObject?,
-    psi: KtLightMethod,
-    givenParent: UElement?
-) : KotlinConstructorUMethod(ktClass, psi, givenParent) {
-    override fun getBodyExpressions(): List<KtExpression> = getInitializers() + super.getBodyExpressions()
 }
 
 open class KotlinUAnnotationMethod(
